feat: re-locate linked resource card
