+++ src/modules/Entities/CreateEntity/CreateEntityAdvanced/components/LinkedResourcesContentCard/LinkedResourcesContentCard.tsx
